// project-factory-frontend/src/app/constant/enums.enum.ts
export enum HttpStatus {
  OK = 200,
  ACCEPTED = 202,
  NO_CONTENT = 204,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  REQUEST_TIMEOUT = 408,
  CONFLICT = 409,
  UNPROCESSABLE_ENTITY = 422,
  INTERNAL_SERVER_ERROR = 500,
  BAD_GATEWAY = 502,
  SERVICE_UNAVAILABLE = 503,
  GATEWAY_TIMEOUT = 504
}

export enum TimeoutConstants {
  TIMEOUT_100 = 100,
  TIMEOUT_300 = 300,
  TIMEOUT_500 = 500,
  TIMEOUT_1000 = 1000,
  TIMEOUT_10000 = 10000,
  TIMEOUT_12000 = 12000
}

export enum ModuleNames {
  PROJECTS = 'projects'
}